pyproject.toml
[project]
name = "cortexpulse-pipeline"
version = "0.1.0"
description = "Minimal news intelligence RAG pipeline using FastAPI, OpenAI, and Qdrant."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "langchain-text-splitters>=0.3.0",
    "openai>=1.50.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.11.0",
    "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
